add session_impl, move charbuff it
