+++ src/posix.cxx
@@ -1,4 +1,4 @@
-#include "ixm/impl.hpp"
+#include "impl.hpp"
 
 #if defined(__ELF__) and __ELF__
   #define SESSION_IMPL_SECTION ".init_array"
